#2 [rename] SocialPlatform 위치 이동
