+++ src/main/java/org/crops/fitserver/auth/facade/AuthFacade.java
@@ -1,6 +1,6 @@
 package org.crops.fitserver.auth.facade;
 
-import org.crops.fitserver.auth.controller.SocialPlatform;
+import org.crops.fitserver.user.domain.SocialPlatform;
 import org.crops.fitserver.auth.facade.dto.TokenResponse;
 
 public interface AuthFacade {
